Added wait for temperature and location and smooth head movement
As it turns out the ender 3 pro will not execute all commands sent via serial connection linearly, echos and temperature/location updates will be send immediately even if the printer is currently waiting on another command such as wait for temperature or move to location. This code requests location and temperature data from the printer to get around this.